Made hook just for useViewerTabShortcuts
